weather_project/requirements.txt
Django
djangorestframework
requests
aiohttp
gunicorn
pytest
